feat(core,arch,sources,pipeline): Phase 10 priority 0 — tail-call promotion + the two bugs that made import names dead on real binaries
Tail-call promotion (ROADMAP Phase 10, priority 0 "CFG fidelity"):
- new `Arch::lift_tail_call` seam (default = `lift`, so ARM64 stays honest
  instead of synthesizing a call it cannot lower; x64 overrides). A `jmp`
  leaving the function used to lift to *nothing*, so a `tail-call` block
  rendered no terminator at all — the call and its returned value were
  silently dropped from the pseudo-C. Now lowered to `Call` + `Return`;
  every style renders `return f(...)`.
- an import thunk (`jmp qword ptr [__imp_X]`) is a tail call, not an
  unrecoverable `ijmp`: the branch is indirect, the callee is known by name.
- `truncate_to_function` now ends a function at a noreturn call (the
  follow-on the 2026-07-22 noreturn pass explicitly left open).

Two bugs found along the way, both of which made the 2026-07-22 noreturn
work dead on every real target while its synthetic unit tests passed:
- `StaticPe` keyed its IAT map by goblin's `Import::rva` (the hint/name
  table entry) instead of `Import::offset` (the IAT slot RVA, despite the
  name). No real `call qword ptr [rip+disp]` ever resolved a name, so the
  noreturn CFG closure, the known-API signature table and thunk detection
  were all inert. Corpus effect: vcruntime140.dll 0 -> 33 functions flagged
  `calls-noreturn`.
- `cfg_cache_key` hashed only target facts, never the analyzer, so
  `.n0x/ir-cache/` served pre-upgrade artifacts and masked the fix. The key
  now carries an analysis fingerprint (crate version + exe mtime) as a
  generation prefix, and `ir_cache::retain_prefix` sweeps other generations
  so the store stays bounded.

Also: an import call renders by name (`kernel32__CloseHandle(...)`) instead
of `(*(uint64_t*)(0x14002a3e8))(...)` — new optional `Callsite.via_slot`
(additive to n0xis.ir.cfg.v1) carries the slot the callee was reached
through, which also lets the known-API signature table fire on IAT calls
for the first time; module names are sanitized into valid C identifiers.

Verified on real PEs, not just fixtures: version.dll thunk ->
`return api_ms_win_core_version_l1_1_1_dll__GetFileVersionInfoSizeW(...)`;
tail flag on 15/400 notepad.exe and 52/400 dxgi.dll functions.

9 new tests (3 ir.rs, 2 x64_lift.rs, 3 render.rs, 1 decomp.rs) plus one in
ir_cache.rs; n0xis-core lib 131->138, n0xis-arch 22->24, workspace green,
clippy clean.

Carries the in-flight WIP already present in the tree (n0xis-frontend
capability registry, CI workflow, boundary check script, docs).

+++ crates/n0xis-core/src/optimize.rs
@@ -221,7 +221,7 @@ fn sign_extend(v: i128, bits: n0xis_arch::Bits) -> i128 {
     if bits == 0 || bits >= 128 {
         return v;
     }
-    let shift = 128 - bits as u32;
+    let shift = 128 - bits;
     (v << shift) >> shift
 }
 
@@ -377,6 +377,17 @@ fn expr_prop_round(blocks: &mut [SsaBlock], delta: &mut Vec<OptDeltaEntry>) -> b
             }
             // Find the single use, restricted to this same block, after `i`,
             // with no Call/Store between.
+            //
+            // KNOWN BUG (surfaced by clippy::never_loop, allow kept so the
+            // gate stays green without a silent behavior change): every arm
+            // below leaves the loop on the FIRST iteration — the trailing
+            // `break` was meant to sit after the loop, not inside it. So this
+            // only ever inspects `j == i + 1`, i.e. propagation happens solely
+            // when the use is the immediately next statement. Moving the
+            // `break` out restores the documented "scan forward to the use"
+            // behavior and changes decompiler output, so it belongs in its own
+            // change with pseudo-C goldens re-checked, not in a CI commit.
+            #[allow(clippy::never_loop)]
             for j in (i + 1)..b.stmts.len() {
                 if stmt_is_barrier(&b.stmts[j].stmt) {
                     i += 1;
